Bug 1736563 - Part 1: Use uint32_t for bfcache status flags r=smaug
Differential Revision: https://phabricator.services.mozilla.com/D129013

UltraBlame original commit: 72c2de8d36d133d2d7a35c0b6b46f201eafe5db2

+++ dom/ipc/ContentChild.cpp
@@ -4455,7 +4455,7 @@ mozilla::ipc::IPCResult ContentChild::RecvCanSavePresentation(
 
   bool canSave = true;
   
-  uint16_t flags = 0;
+  uint32_t flags = 0;
   BrowsingContext* browsingContext = aTopLevelContext.get();
   browsingContext->PreOrderWalk([&](BrowsingContext* aContext) {
     Document* doc = aContext->GetDocument();
